chore(weave): move builtin object schemas to weave/shared
The builtin object classes (Leaderboard, SavedView, AnnotationSpec, Provider,
LLMStructuredCompletionModel, ...) are pydantic schemas shared by the client
(which re-exports and subclasses them) and the trace server (which validates
against them). Move the package out of weave/trace_server/interface into
weave/shared so client code can stop importing weave.trace_server:

- weave/trace_server/interface/builtin_object_classes/** ->
  weave/shared/builtin_object_classes/** (including generated/ schemas json)
- weave/trace_server/helpers/url_safety.py -> weave/shared/url_safety.py
  (imported by the provider schema; also used by server image_completion)

Old import paths keep working via shims; the shims pair a star re-export
(for static analyzers) with a PEP 562 module __getattr__ fallback so names a
module's __all__ hides (e.g. TestOnlyNestedBaseModel) still resolve. The
schema generation script and dev docs now point at the new location; the
generated json is byte-identical after regeneration. Server-side imports are
left on the shims on purpose.
