+++ dev_docs/BuiltinObjectClasses.md
@@ -80,7 +80,7 @@ While many Weave Objects are free-form and user-defined, there is often a need f
 
 Here's how to define and use a validated base object:
 
-1. **Define your schema** (in `weave/trace_server/interface/builtin_object_classes/your_schema.py`):
+1. **Define your schema** (in `weave/shared/builtin_object_classes/your_schema.py`):
 
 ```python
 from pydantic import BaseModel
@@ -160,10 +160,10 @@ Run `make synchronize-base-object-schemas` to ensure the frontend TypeScript typ
 
 ### Architecture Flow
 
-1. Define your schema in a python file in the `weave/trace_server/interface/builtin_object_classes/test_only_example.py` directory. See `weave/trace_server/interface/builtin_object_classes/test_only_example.py` as an example.
-2. Make sure to register your schemas in `weave/trace_server/interface/builtin_object_classes/builtin_object_registry.py` by calling `register_base_object`.
+1. Define your schema in a python file in the `weave/shared/builtin_object_classes/test_only_example.py` directory. See `weave/shared/builtin_object_classes/test_only_example.py` as an example.
+2. Make sure to register your schemas in `weave/shared/builtin_object_classes/builtin_object_registry.py` by calling `register_base_object`.
 3. Run `make synchronize-base-object-schemas` to generate the frontend types.
-   - The first step (`make generate_base_object_schemas`) will run `scripts/generate_base_object_schemas.py` to generate a JSON schema in `weave/trace_server/interface/builtin_object_classes/generated/generated_builtin_object_class_schemas.json`.
+   - The first step (`make generate_base_object_schemas`) will run `scripts/generate_base_object_schemas.py` to generate a JSON schema in `weave/shared/builtin_object_classes/generated/generated_builtin_object_class_schemas.json`.
    - The second step (yarn `generate-schemas`) will read this file and use it to generate the frontend types located in `frontends/weave/src/components/PagePanelComponents/Home/Browse3/pages/wfReactInterface/generatedBuiltinObjectClasses.zod.ts`.
 4. Now, each use case uses different parts:
    1. `Python Writing`. Users can directly import these classes and use them as normal Pydantic models, which get published with `weave.publish`. The python client correct builds the requisite payload.
