+++ weave/flow/annotation_spec.py
@@ -1,4 +1,4 @@
-from weave.trace_server.interface.builtin_object_classes import annotation_spec
+from weave.shared.builtin_object_classes import annotation_spec
 
 # Re-export:
 AnnotationSpec = annotation_spec.AnnotationSpec
