+++ weave/flow/leaderboard.py
@@ -3,9 +3,9 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Any
 
+from weave.shared.builtin_object_classes import leaderboard
 from weave.trace.refs import OpRef
 from weave.trace.weave_client import WeaveClient, get_ref
-from weave.trace_server.interface.builtin_object_classes import leaderboard
 from weave.trace_server.trace_server_interface import CallsFilter
 from weave.utils.project_id import from_project_id
