+++ weave/scorers/llm_as_a_judge_scorer.py
@@ -4,13 +4,13 @@ from pydantic import AliasChoices, ConfigDict, Field, field_validator
 
 from weave.flow.scorer import Scorer
 from weave.prompt.prompt import MessagesPrompt
+from weave.shared.builtin_object_classes.llm_structured_model import (
+    LLMStructuredCompletionModel,
+)
 from weave.trace.context.weave_client_context import get_weave_client
 from weave.trace.objectify import maybe_objectify, register_object
 from weave.trace.op import op
 from weave.trace.vals import make_trace_obj
-from weave.trace_server.interface.builtin_object_classes.llm_structured_model import (
-    LLMStructuredCompletionModel,
-)
 
 
 @register_object
