+++ weave/shared/builtin_object_classes/leaderboard.py
@@ -1,0 +1,14 @@
+from pydantic import BaseModel
+
+from weave.shared.builtin_object_classes import base_object_def
+
+
+class LeaderboardColumn(BaseModel):
+    evaluation_object_ref: base_object_def.RefStr
+    scorer_name: str
+    summary_metric_path: str
+    should_minimize: bool | None = None
+
+
+class Leaderboard(base_object_def.BaseObject):
+    columns: list[LeaderboardColumn]
