+++ weave/trace/api.py
@@ -11,6 +11,7 @@ from typing import Any, cast
 # TODO: type_handlers is imported here to trigger registration of the image serializer.
 # There is probably a better place for this, but including here for now to get the fix in.
 from weave import type_handlers  # noqa: F401
+from weave.shared.builtin_object_classes import leaderboard
 from weave.shared.ids import generate_id
 from weave.trace import urls, weave_client, weave_init
 from weave.trace.autopatch import AutopatchSettings
@@ -28,7 +29,6 @@ from weave.trace.settings import (
 )
 from weave.trace.table import Table
 from weave.trace.view_utils import set_call_view
-from weave.trace_server.interface.builtin_object_classes import leaderboard
 from weave.trace_server_bindings.link_asset_to_registry import LinkAssetToRegistryRes
 from weave.type_wrappers.Content.content import Content
