+++ weave/trace/base_objects.py
@@ -1,4 +1,4 @@
-from weave.trace_server.interface.builtin_object_classes.builtin_object_registry import (
+from weave.shared.builtin_object_classes.builtin_object_registry import (
     BUILTIN_OBJECT_REGISTRY,
     AlertSpec,
     AnnotationSpec,
